LPS-157816 Extract FieldTypeSelector
